raid/requirements.txt
pyyaml>=5.4.1
python-evtx>=0.7.4